LPD-17452 Move common constants and types to `utils` dir
