+++ modules/apps/frontend-data-set/frontend-data-set-views-web/src/main/resources/META-INF/resources/js/components/OrderableTable.tsx
@@ -16,7 +16,7 @@ import React, {useEffect, useRef, useState} from 'react';
 import {DndProvider, useDrag, useDrop} from 'react-dnd';
 import {HTML5Backend} from 'react-dnd-html5-backend';
 
-import {FUZZY_OPTIONS} from '../Constants';
+import {FUZZY_OPTIONS} from '../utils/constants';
 import Search from './Search';
 
 import '../../css/OrderableTable.scss';
