+++ modules/apps/frontend-data-set/frontend-data-set-views-web/src/main/resources/META-INF/resources/js/components/modal_content/SelectionFilter.tsx
@@ -14,7 +14,7 @@ import classNames from 'classnames';
 import fuzzy from 'fuzzy';
 import React, {useState} from 'react';
 
-import {IPickList} from '../../types';
+import {IPickList} from '../../utils/types';
 import CheckboxMultiSelect from '../CheckboxMultiSelect';
 
 function Header() {
