+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/api.d.ts
@@ -4,7 +4,7 @@
  */
 
 import {FDSViewType} from './FDSViews';
-import {IField, IPickList} from './types';
+import {IField, IPickList} from './utils/types';
 export declare function getFields(fdsView: FDSViewType): Promise<IField[]>;
 export declare function getAllPicklists(
 	page?: number,
