+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/components/modal_content/SelectionFilter.d.ts
@@ -5,7 +5,7 @@
 
 /// <reference types="react" />
 
-import {IPickList} from '../../types';
+import {IPickList} from '../../utils/types';
 declare function Header(): JSX.Element;
 interface IBodyProps {
 	includeMode: string;
