+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/fds_view/visualization_modes/table/modal_content/AddFieldsModalContent.d.ts
@@ -6,7 +6,7 @@
 import {TreeView} from '@clayui/core';
 import {ComponentProps} from 'react';
 import {FDSViewType} from '../../../../FDSViews';
-import {IFDSField} from '../Table';
+import {IFDSField} from '../../../../utils/types';
 declare const AddFieldsModalContent: ({
 	closeModal,
 	fdsView,
