+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/utils/types.d.ts
@@ -3,27 +3,23 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-type LocalizedValue<T> = Liferay.Language.LocalizedValue<T>;
-
-export enum EFilterType {
+declare type LocalizedValue<T> = Liferay.Language.LocalizedValue<T>;
+export declare enum EFilterType {
 	CLIENT_EXTENSION = 'CLIENT_EXTENSION',
 	DATE_RANGE = 'DATE_RANGE',
 	SELECTION = 'SELECTION',
 }
-
-export enum EFieldFormat {
+export declare enum EFieldFormat {
 	DATE = 'date',
 	DATE_TIME = 'date-time',
 	INT64 = 'int64',
 }
-
-export enum EFieldType {
+export declare enum EFieldType {
 	ARRAY = 'array',
 	INTEGER = 'integer',
 	OBJECT = 'object',
 	STRING = 'string',
 }
-
 export interface IField {
 	children?: Array<IField>;
 	format?: EFieldFormat;
@@ -31,10 +27,21 @@ export interface IField {
 	label?: string;
 	name: string;
 	selected?: boolean;
-	type: string;
+	type?: string;
 	visible?: boolean;
 }
-
+export interface IFDSField {
+	contextPath: string;
+	externalReferenceCode: string;
+	id: number;
+	label: string;
+	label_i18n: LocalizedValue<string>;
+	name: string;
+	renderer: string;
+	rendererLabel?: string;
+	sortable: boolean;
+	type: string;
+}
 export interface IFilter {
 	fieldName: string;
 	filterType?: EFilterType;
@@ -43,23 +50,19 @@ export interface IFilter {
 	label_i18n: LocalizedValue<string>;
 	type: string;
 }
-
 export interface IClientExtensionFilter extends IFilter {
 	fdsFilterClientExtensionERC: string;
 }
-
 export interface IDateFilter extends IFilter {
 	from: string;
 	to: string;
 }
-
 export interface ISelectionFilter extends IFilter {
 	include: boolean;
 	listTypeDefinitionERC: string;
 	multiple: boolean;
 	preselectedValues: string;
 }
-
 export interface IPickList {
 	externalReferenceCode: string;
 	id: string;
@@ -69,7 +72,6 @@ export interface IPickList {
 		[key: string]: string;
 	};
 }
-
 export interface IListTypeEntry {
 	externalReferenceCode: string;
 	id: number;
@@ -79,3 +81,4 @@ export interface IListTypeEntry {
 		[key: string]: string;
 	};
 }
+export {};
